+++ src/components/todoList.tsx
@@ -1,8 +1,8 @@
-import React, { ChangeEvent, KeyboardEvent, MouseEvent, useState } from 'react';
+import React from 'react';
 import { TodoListHeader } from './TodoListHeader';
 import { AddItemForm } from './AddItemForm';
 import { ButtonsFilter } from './ButtonsFilter';
-import classes from './TodoList.module.css';
+import { TodoListTasks } from './TodoListTasks';
 
 export type TaskType = {
   id: string;
@@ -24,25 +24,11 @@ type PropsType = {
 export function TodoList(props: PropsType) {
   const { title, tasks, filterValue, removeTask, changeFilter, addTask, changeTaskStatus } = props;
 
-  const mappedTasks = tasks.map((task) => {
-    const onRemoveHandler = (): void => removeTask(task.id);
-    const onStatusHandler = (e: ChangeEvent<HTMLInputElement>): void => {
-      changeTaskStatus(task.id, e.currentTarget.checked);
-    };
-    return (
-      <li key={task.id}>
-        <input type='checkbox' checked={task.isDone} onChange={onStatusHandler} />
-        <span className={task.isDone ? classes.task_completed : ''}>{task.title}</span>
-        <button onClick={onRemoveHandler}>x</button>
-      </li>
-    );
-  });
-
   return (
     <div>
       <TodoListHeader title={title} />
       <AddItemForm addTask={addTask} />
-      <ul>{mappedTasks}</ul>
+      <TodoListTasks tasks={tasks} removeTask={removeTask} changeTaskStatus={changeTaskStatus} />
       <ButtonsFilter filterValue={filterValue} changeFilter={changeFilter} />
     </div>
   );
